requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.19
django-stdimage==5.3.0
gunicorn==20.0.4
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
